requirements.txt
numpy
pandas
pillow
scikit-image
matplotlib
requests
